Getting document frequency of a term
Added get_document_frequency() and get_idf_weight methods() to the BM25 class.
They both take in a term as a parameter.

Also tweaked the Query classes to not be initialized with the terms (i.e. the query).
We pass it in as an argument to the execute() method.
That way, we don't have to create a new Query object every time we want to conduct a query.

Set a number_of_documents instance variable in the Reuters class, so that I can easily access it.
Will be used to define the total number of documents (N) in BM25.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>